refactor: usePlanetColor hook으로 분리
